+++ src/app/editor-toolbar/editor-toolbar.models.ts
@@ -31,7 +31,3 @@ export function getDecIncValue(tool: DrawingTools): number {
     
     return 0
 }
-
-export function isPlayerCommand(tool: DrawingTools): boolean {
-    return Object.values([DrawingTools.RUN, DrawingTools.STEP, DrawingTools.STOP, DrawingTools.PAUSE]).includes(tool)
-}
